requirements.txt
pyttsx3
opencv-python
google-cloud-vision
cohere